refactor: イベントのバリデーションのコードを別ファイルに移動
