+++ app/javascript/components/EventForm.js
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import { isEmptyObject, validateEvent } from '../helpers/helpers';
 
 const EventForm = () => {
   const [event, setEvent] = useState({
@@ -19,34 +20,6 @@ const EventForm = () => {
 
     setEvent({ ...event, [name]: value });
   };
-
-  const validateEvent = () => {
-    const errors = {};
-
-    if (event.event_type === '') {
-      errors.event_type = 'You must enter an event type';
-    }
-
-    if (event.event_date === '') {
-      errors.event_date = 'You must enter a valid date';
-    }
-
-    if (event.title === '') {
-      errors.title = 'You must enter a title';
-    }
-
-    if (event.speaker === '') {
-      errors.speaker = 'You must enter at least one speaker';
-    }
-
-    if (event.host === '') {
-      errors.host = 'You must enter at least one host';
-    }
-
-    return errors;
-  };
-
-  const isEmptyObject = (obj) => Object.keys(obj).length === 0;
 
   const renderErrors = () => {
     if (isEmptyObject(formErrors)) {
